--- eq-author/src/components/AnswerContent/AnswerProperties/Answers/index.js
@@ -1,5 +1,0 @@
-export { default as Decimal } from "../Decimal";
-export { default as DateProperties } from "./DateProperties";
-export { default as UnitProperties } from "./UnitProperties";
-export { default as DurationProperties } from "./DurationProperties";
-export { default as TextFieldProperties } from "./TextFieldProperties";
