+++ eq-author/src/components/AnswerContent/AnswerProperties/CheckboxProperties.js
