+++ eq-author/src/components/AnswerContent/AnswerProperties/DurationProperties.js
