+++ eq-author/src/components/AnswerContent/AnswerProperties/DurationProperties.test.js
