+++ eq-author/src/components/AnswerContent/AnswerProperties/NumberProperties.test.js
