+++ eq-author/src/components/AnswerContent/Decimal/index.js
