+++ eq-author/src/components/AnswerContent/Decimal/index.test.js
